+++ tests/unit/test_routes.py
@@ -148,6 +148,9 @@ def test_routes(warehouse):
             domain=warehouse,
         ),
         pretend.call(
+            "accounts.reauthenticate", "/accounts/reauthenticate/", domain=warehouse
+        ),
+        pretend.call(
             "accounts.recovery-code", "/account/recovery-code/", domain=warehouse
         ),
         pretend.call("accounts.logout", "/account/logout/", domain=warehouse),
@@ -300,7 +303,6 @@ def test_routes(warehouse):
             domain=warehouse,
         ),
         pretend.call("packaging.file", "https://files.example.com/packages/{path}"),
-        pretend.call("reauthenticate", "/reauthenticate/", domain=warehouse),
         pretend.call("ses.hook", "/_/ses-hook/", domain=warehouse),
         pretend.call("rss.updates", "/rss/updates.xml", domain=warehouse),
         pretend.call("rss.packages", "/rss/packages.xml", domain=warehouse),
