+++ warehouse/templates/pages/help.html
@@ -475,7 +475,7 @@
         <h3 id="sensitiveactions">{{ sensitiveactions() }}</h3>
 
         {% trans %}
-          <p>PyPI asks you to confirm your password before you want to perform a <i>sensitive action</i>. Sensitive actions include things like adding or removing maintainers, deleting distributions, generating API tokens, and setting up 2FA.</p>
+          <p>PyPI asks you to confirm your password before you want to perform a <i>sensitive action</i>. Sensitive actions include things like adding or removing maintainers, deleting distributions, generating API tokens, and setting up two-factor authentication.</p>
           <p>You'll only have to re-confirm your password if it's been more than an hour since you last confirmed it.</p>
           <p><strong>We strongly recommend you only perform such actions on your personal, password-protected computer.</strong></p>
