+++ warehouse/views.py
@@ -12,7 +12,6 @@
 
 
 import collections
-import json
 
 import elasticsearch
 
@@ -42,7 +41,6 @@ from sqlalchemy.sql import exists
 from trove_classifiers import classifiers, deprecated_classifiers
 
 from warehouse.accounts import REDIRECT_FIELD_NAME
-from warehouse.accounts.interfaces import IUserService
 from warehouse.accounts.models import User
 from warehouse.cache.http import add_vary, cache_control
 from warehouse.cache.origin import origin_cache
@@ -50,7 +48,6 @@ from warehouse.classifiers.models import Classifier
 from warehouse.db import DatabaseNotAvailable
 from warehouse.forms import SetLocaleForm
 from warehouse.i18n import LOCALE_ATTR
-from warehouse.manage.forms import ReAuthenticateForm
 from warehouse.metrics import IMetricsService
 from warehouse.packaging.models import File, Project, Release, release_classifiers
 from warehouse.search.queries import SEARCH_FILTER_ORDER, get_es_query
@@ -462,38 +459,3 @@ def force_status(request):
         raise exception_response(int(request.matchdict["status"]))
     except KeyError:
         raise exception_response(404) from None
-
-
-@view_config(
-    route_name="reauthenticate",
-    uses_session=True,
-    require_csrf=True,
-    require_methods=False,
-    has_translations=True,
-)
-def reauthenticate(request, _form_class=ReAuthenticateForm):
-    user_service = request.find_service(IUserService, context=None)
-
-    form = _form_class(
-        request.POST,
-        request=request,
-        user_service=user_service,
-        check_password_metrics_tags=[
-            "method:reauth",
-            "auth_method:reauthenticate_form",
-        ],
-    )
-
-    if form.next_route.data and form.next_route_matchdict.data:
-        redirect_to = request.route_path(
-            form.next_route.data, **json.loads(form.next_route_matchdict.data)
-        )
-    else:
-        redirect_to = request.route_path("manage.projects")
-
-    resp = HTTPSeeOther(redirect_to)
-
-    if request.method == "POST" and form.validate():
-        request.session.record_auth_timestamp()
-
-    return resp
